+++ packages/editor/src/store/test/selectors.jsx
@@ -95,8 +95,8 @@ selectorNames.forEach( ( name ) => {
 				return state.getCurrentUser && state.getCurrentUser();
 			},
 
-			hasFetchedAutosaves() {
-				return state.hasFetchedAutosaves && state.hasFetchedAutosaves();
+			hasFetchedAutosave() {
+				return state.hasFetchedAutosave && state.hasFetchedAutosave();
 			},
 
 			getAutosave() {
@@ -1406,7 +1406,7 @@ describe( 'selectors', () => {
 				postAutosavingLock: {},
 				saving: {},
 				getCurrentUser() {},
-				hasFetchedAutosaves() {
+				hasFetchedAutosave() {
 					return false;
 				},
 				getAutosave() {
@@ -1436,7 +1436,7 @@ describe( 'selectors', () => {
 				postAutosavingLock: {},
 				saving: {},
 				getCurrentUser() {},
-				hasFetchedAutosaves() {
+				hasFetchedAutosave() {
 					return true;
 				},
 				getAutosave() {
@@ -1465,7 +1465,7 @@ describe( 'selectors', () => {
 				},
 				saving: {},
 				getCurrentUser() {},
-				hasFetchedAutosaves() {
+				hasFetchedAutosave() {
 					return true;
 				},
 				getAutosave() {},
@@ -1493,7 +1493,7 @@ describe( 'selectors', () => {
 				},
 				saving: {},
 				getCurrentUser() {},
-				hasFetchedAutosaves() {
+				hasFetchedAutosave() {
 					return true;
 				},
 				getAutosave() {
@@ -1523,7 +1523,7 @@ describe( 'selectors', () => {
 				},
 				saving: {},
 				getCurrentUser() {},
-				hasFetchedAutosaves() {
+				hasFetchedAutosave() {
 					return true;
 				},
 				getAutosave() {
@@ -1562,7 +1562,7 @@ describe( 'selectors', () => {
 						},
 						saving: {},
 						getCurrentUser() {},
-						hasFetchedAutosaves() {
+						hasFetchedAutosave() {
 							return true;
 						},
 						getAutosave() {
@@ -1605,7 +1605,7 @@ describe( 'selectors', () => {
 				},
 				saving: {},
 				getCurrentUser() {},
-				hasFetchedAutosaves() {
+				hasFetchedAutosave() {
 					return true;
 				},
 				getAutosave() {},